refactor: migrate NPM to Yarn and add TypeScript and Prettier
