+++ .github/CODEOWNERS
@@ -6,6 +6,7 @@
 .github/pull_request_template.md    @Ir1d
 .github/workflows/build-pdf.yml	@Enter-tainer
 .github/workflows/build.yml @billchenchina
+.github/workflows/build-authors-cache.yml @Menci
 .github/workflows/celebration.yml   @ouuan
 .github/workflows/pr-stale.yml	@CoelacanthusHex
 .github/ISSUE_TEMPLATE/	@Ir1d
@@ -14,12 +15,14 @@
 Dockerfile  @CoelacanthusHex
 docs/intro/docker-deploy.md @CoelacanthusHex
 scripts/    @Ir1d @billchenchina
+.prettierrc   @Menci
 .clang-format   @Ir1d
 .remarkignore   @Enter-tainer
 .remarkrc   @Ir1d @Enter-tainer
 CODE_OF_CONDUCT.md  @Ir1d
 README.md   @Ir1d
-gulpfile.js @Ir1d @Enter-tainer
+gulpfile.cjs @Ir1d @Enter-tainer
 package.json    @Ir1d @Enter-tainer
 requirements.txt    @Ir1d
 runtime.txt @Irld
+tsconfig.json @Menci
