+++ .github/workflows/build.yml
@@ -31,23 +31,25 @@ jobs:
         id: cache-node
         with:
           path: node_modules
-          key: ${{ runner.os }}-node-${{ hashFiles('package-lock.json') }}
+          key: ${{ runner.os }}-node-${{ hashFiles('yarn.lock') }}
           restore-keys: ${{ runner.os }}-node-
       - name: Install Node.js dependencies
         if: steps.cache-node.outputs.cache-hit != 'true'
-        run: npm install
+        run: yarn --frozen-lockfile --production
       - name: Page Build
         run: |
           chmod +x ./scripts/build.sh && ./scripts/build.sh
           mkdocs build -v
-      - name: Render git history info
-        run: ./scripts/render_git_history_info.sh
+      - name: Render commits info
+        run: ./scripts/render-commits-info.sh
         env:
           GITHUB_TOKEN: ${{ secrets.GITHUB_TOKEN }}
       - name: Render math
-        run: node --max_old_space_size=3072 ./scripts/render_math.js
+        run: yarn ts-node-esm ./scripts/render_math.ts
+        env:
+          NODE_OPTIONS: --max_old_space_size=3072
       - name: Gulp minify
-        run: npx gulp minify
+        run: yarn gulp -f gulpfile.cjs minify
       - name: Generate redirects
         run: python scripts/gen_redirect.py
       - name: Deploy to gh-pages
