+++ .github/workflows/check-scripts.yml
@@ -1,0 +1,38 @@
+name: Check Scripts
+
+on:
+  pull_request:
+    branches:
+      - master
+    paths:
+      - scripts/**
+      - package.json
+      - yarn.lock
+      - .prettierrc
+      - tsconfig.json
+  workflow_dispatch:
+
+jobs:
+  build:
+    name: Check Scripts
+    runs-on: ubuntu-latest
+    steps:
+      - uses: actions/checkout@v2
+      - uses: actions/setup-node@v2
+        with:
+          node-version: '18.x'
+          check-latest: true
+      - name: Cache Node.js dependencies
+        uses: actions/cache@v2
+        id: cache-node
+        with:
+          path: node_modules
+          key: ${{ runner.os }}-node-${{ hashFiles('yarn.lock') }}
+          restore-keys: ${{ runner.os }}-node-
+      - name: Install Node.js dependencies
+        if: steps.cache-node.outputs.cache-hit != 'true'
+        run: yarn --frozen-lockfile
+      - name: Check Prettier Format
+        run: yarn scripts:format:check
+      - name: Check TypeScript Compile
+        run: yarn tsc
