+++ .prettierrc
@@ -1,0 +1,6 @@
+{
+  "endOfLine": "lf",
+  "printWidth": 120,
+  "arrowParens": "avoid",
+  "trailingComma": "none"
+}
