+++ scripts/authors-cache.ts
@@ -1,0 +1,6 @@
+export type AuthorUserMap = Record<string, { name: string; githubUsername: string }>;
+
+export interface AuthorsCache {
+  latestCommitTime: string;
+  userMap: AuthorUserMap;
+}
